backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.34.2
psycopg2-binary==2.9.9
aiosqlite==0.20.0
greenlet==3.1.1
